minor improvements

+++ core/src/dma_channel.rs
@@ -1,6 +1,6 @@
 #![allow(non_camel_case_types)]
 
-use crate::bus::{Bus, ChunkSize, MemoryRegion, CartridgeType};
+use crate::bus::{Bus, CartridgeType, ChunkSize, MemoryRegion};
 
 #[derive(Clone, Copy, PartialEq)]
 pub enum TimingMode {
@@ -13,10 +13,10 @@ pub enum TimingMode {
 #[derive(Clone)]
 pub struct DMA_Channel {
     channel_no: usize,
-    pub src_addr: usize,
-    pub dest_addr: usize,
-    src_increment: usize,  // -1, 0, 1.
-    dest_increment: usize, // -1, 0, 1.
+    pub src_addr: u32,
+    pub dest_addr: u32,
+    src_increment: u32,  // -1, 0, 1.
+    dest_increment: u32, // -1, 0, 1.
     num_transfers: u16,
     chunk_size: ChunkSize,
     pub timing_mode: TimingMode,
@@ -45,8 +45,8 @@ impl DMA_Channel {
     }
 
     pub fn new_enabled(channel_no: usize, bus: &mut Bus) -> DMA_Channel {
-        let src_addr = bus.read_word_raw(0xb0 + 12 * channel_no, MemoryRegion::IO) as usize;
-        let dest_addr = bus.read_word_raw(0xb4 + 12 * channel_no, MemoryRegion::IO) as usize;
+        let src_addr = bus.read_word_raw(0xb0 + 12 * channel_no, MemoryRegion::IO);
+        let dest_addr = bus.read_word_raw(0xb4 + 12 * channel_no, MemoryRegion::IO);
         let dma_cnt = bus.read_word_raw(0xb8 + 12 * channel_no, MemoryRegion::IO);
         let mut num_transfers = dma_cnt as u16;
         let timing_mode = match (dma_cnt >> 0x1c) & 0b11 {
@@ -110,7 +110,9 @@ impl DMA_Channel {
                             }
                             // sound FIFO mode
                             1 | 2 => {
-                                bus.apu.direct_sound_fifo[(self.dest_addr - 0x040000a0) >> 2].len()
+                                bus.apu.direct_sound_fifo
+                                    [(self.dest_addr as usize - 0x040000a0) >> 2]
+                                    .len()
                                     <= 16
                             }
                             // video transfer mode
@@ -145,8 +147,7 @@ impl DMA_Channel {
                 _ => dma_cnt as u16,
             };
             if self.repeat_reset_dest {
-                self.dest_addr =
-                    bus.read_word_raw(0xb4 + 12 * self.channel_no, MemoryRegion::IO) as usize;
+                self.dest_addr = bus.read_word_raw(0xb4 + 12 * self.channel_no, MemoryRegion::IO);
             }
         }
 
@@ -189,8 +190,7 @@ impl DMA_Channel {
             assert!(self.num_transfers == 4);
             assert!(self.dest_addr == 0x040000a0 || self.dest_addr == 0x040000a4);
             assert!(self.check_is_active(bus));
-        }
-        else{
+        } else {
             panic!("video transfer DMA not implemented");
         }
 
@@ -202,14 +202,21 @@ impl DMA_Channel {
         if self.channel_no != 1 && self.channel_no != 2 {
             //println!("dest: {:#x}, channel_no: {}", self.dest_addr, self.channel_no);
         }
-        if self.channel_no == 3{
+        if self.channel_no == 3 {
             //println!("dma channel 3, src addr: {:#x}, dest addr: {:#x}", self.src_addr, self.dest_addr);
         }
         //
-        if self.channel_no == 3 && ((self.src_addr >= 0xd000000 && self.src_addr <= 0xdffffff ) || (self.dest_addr >= 0xd000000 && self.dest_addr <= 0xdffffff )) 
-            && (bus.cartridge_type == CartridgeType::Eeprom512 || bus.cartridge_type == CartridgeType::Eeprom8192){
+        if self.channel_no == 3
+            && ((self.src_addr >= 0xd000000 && self.src_addr <= 0xdffffff)
+                || (self.dest_addr >= 0xd000000 && self.dest_addr <= 0xdffffff))
+            && (bus.cartridge_type == CartridgeType::Eeprom512
+                || bus.cartridge_type == CartridgeType::Eeprom8192)
+        {
             //println!("chunksize: {}, src_inc: {}, dest_inc: {}", self.chunk_size as u32, self.src_increment as i32, self.dest_increment as i32);
-            if self.chunk_size == ChunkSize::Halfword && self.src_increment == 1 && self.dest_increment == 1 {
+            if self.chunk_size == ChunkSize::Halfword
+                && self.src_increment == 1
+                && self.dest_increment == 1
+            {
                 //if self.dest_addr >= 0xd000000 && self.dest_addr <= 0xdffffff && self.src_addr >= 0xd000000 && self.src_addr <= 0xdffffff{
                 //    println!("eeprom src and dest both in eeprom region");
                 //}
@@ -221,8 +228,8 @@ impl DMA_Channel {
                     let mut sram_addr = 0;
                     let mut j = 0;
                     let mut is_read = false;
-                    for i in 0..self.num_transfers{
-                        let data = bus.read_halfword(self.src_addr);
+                    for i in 0..self.num_transfers {
+                        let data = bus.read_halfword(self.src_addr as usize);
                         res <<= 1;
                         res |= data as u64 & 1;
                         j += 1;
@@ -231,20 +238,22 @@ impl DMA_Channel {
                                 is_read = false;
                                 bus.eeprom_is_read = false;
                                 //println!("eeprom write");
-                            }
-                            else if res == 0b11 {
+                            } else if res == 0b11 {
                                 is_read = true;
                                 bus.eeprom_is_read = true;
                                 //println!("eeprom read set addr");
-                            } 
-                            else{
-                                println!("DMA channel 3 EEPROM no matching bits, res: {:#05b}", res);
+                            } else {
+                                println!(
+                                    "DMA channel 3 EEPROM no matching bits, res: {:#05b}",
+                                    res
+                                );
                                 break;
                             }
                             j = 0;
                             res = 0;
-                        }
-                        else if (i == 7 && bus.cartridge_type == CartridgeType::Eeprom512) || (i == 15 && bus.cartridge_type == CartridgeType::Eeprom8192){
+                        } else if (i == 7 && bus.cartridge_type == CartridgeType::Eeprom512)
+                            || (i == 15 && bus.cartridge_type == CartridgeType::Eeprom8192)
+                        {
                             //assert!(res < 0x400);
                             sram_addr = res << 3;
                             j = 0;
@@ -253,13 +262,20 @@ impl DMA_Channel {
                                 //println!("EEPROM setting read mem addr");
                                 bus.eeprom_read_offset = sram_addr as usize;
                             }
-                        }
-                        else if !is_read && j == 64{
+                        } else if !is_read && j == 64 {
                             //println!("EEPROM writing to memory, i: {}, num_transfers: {}", i, self.num_transfers);
                             j = 0;
                             let base_addr = sram_addr;
-                            bus.store_word_raw(base_addr as usize, MemoryRegion::CartridgeSram, res as u32);
-                            bus.store_word_raw(base_addr as usize + 4, MemoryRegion::CartridgeSram, (res >> 32) as u32);
+                            bus.store_word_raw(
+                                base_addr as usize,
+                                MemoryRegion::CartridgeSram,
+                                res as u32,
+                            );
+                            bus.store_word_raw(
+                                base_addr as usize + 4,
+                                MemoryRegion::CartridgeSram,
+                                (res >> 32) as u32,
+                            );
                             //println!("write res: {:#18x}", res);
                             //println!("write base addr: {:#x}", base_addr);
                             bus.eeprom_write_successful = true;
@@ -267,63 +283,62 @@ impl DMA_Channel {
 
                         self.src_addr += 2;
                         self.dest_addr += 2;
-                    }  
+                    }
                 }
                 // EEPROM read
-                else if bus.eeprom_is_read{
+                else if bus.eeprom_is_read {
                     //println!("eeprom read");
                     //bus.eeprom_is_read = false;
                     let mut j = 0;
                     let base_addr = bus.eeprom_read_offset;
                     //println!("read base addr: {:#x}", base_addr);
-                    let res = bus.read_word_raw(base_addr as usize, MemoryRegion::CartridgeSram) as u64 +
-                        ((bus.read_word_raw(base_addr as usize + 4, MemoryRegion::CartridgeSram) as u64) << 32);
+                    let res = bus.read_word_raw(base_addr as usize, MemoryRegion::CartridgeSram)
+                        as u64
+                        + ((bus.read_word_raw(base_addr as usize + 4, MemoryRegion::CartridgeSram)
+                            as u64)
+                            << 32);
                     //println!("read res: {:#18x}", res);
-                    for i in 0..self.num_transfers{
+                    for i in 0..self.num_transfers {
                         j += 1;
                         let mut data = 0;
                         if i == 3 {
                             j = 0;
-                        }
-                        else if i > 3 && j <= 64 {
+                        } else if i > 3 && j <= 64 {
                             //println!("j: {}", j);
-                            data = ((res >> (64-j)) & 1) as u16;
+                            data = ((res >> (64 - j)) & 1) as u16;
                         }
 
-                        bus.store_halfword(self.dest_addr, data);
+                        bus.store_halfword(self.dest_addr as usize, data);
 
                         self.src_addr += 2;
                         self.dest_addr += 2;
-                    }  
+                    }
                 }
-            }
-            else{
+            } else {
                 println!("fatal error: eeprom DMA 3 has invalid config. chunksize: {}, src_inc: {}, dest_inc: {}", self.chunk_size as u32, self.src_increment as i32, self.dest_increment as i32);
             }
-        }
-        else if self.timing_mode != TimingMode::FIFO{
+        } else if self.timing_mode != TimingMode::FIFO {
             for _ in 0..self.num_transfers {
                 //println!("dest: {:#x}, src: {:#x}, data: {:#010x}", self.dest_addr, self.src_addr, bus.read_word(self.src_addr));
                 match self.chunk_size {
                     ChunkSize::Halfword => {
-                        let data = bus.read_halfword(self.src_addr);
-                        bus.store_halfword(self.dest_addr, data);
+                        let data = bus.read_halfword(self.src_addr as usize);
+                        bus.store_halfword(self.dest_addr as usize, data);
                     }
                     ChunkSize::Word => {
-                        let data = bus.read_word(self.src_addr);
-                        bus.store_word(self.dest_addr, data);
+                        let data = bus.read_word(self.src_addr as usize);
+                        bus.store_word(self.dest_addr as usize, data);
                     }
                     _ => {
                         println!("DMA chunk size must be Word or Halfword");
                     }
                 };
-                self.src_addr += self.src_increment * self.chunk_size as usize;
-                self.dest_addr += self.dest_increment * self.chunk_size as usize;
+                self.src_addr += self.src_increment * self.chunk_size as u32;
+                self.dest_addr += self.dest_increment * self.chunk_size as u32;
             }
-        }
-        else{
-            let channel_num = (self.dest_addr- 0x040000a0) >> 2;
-            for _ in 0..self.num_transfers{
+        } else {
+            let channel_num = (self.dest_addr as usize - 0x040000a0) >> 2;
+            for _ in 0..self.num_transfers {
                 /*if self.dest_addr == 0x040000a0{
                     println!("src addr:     {:#x}", self.src_addr);
                 }
@@ -332,23 +347,28 @@ impl DMA_Channel {
                 }*/
                 match self.chunk_size {
                     ChunkSize::Word => {
-                        let word = bus.read_word(self.src_addr);
+                        let word = bus.read_word(self.src_addr as usize);
                         bus.apu.direct_sound_fifo[channel_num].push_back((word & 0b11111111) as i8);
-                        bus.apu.direct_sound_fifo[channel_num].push_back(((word >> 8) & 0b11111111) as i8);
-                        bus.apu.direct_sound_fifo[channel_num].push_back(((word >> 16) & 0b11111111) as i8);
-                        bus.apu.direct_sound_fifo[channel_num].push_back(((word >> 24) & 0b11111111) as i8);
+                        bus.apu.direct_sound_fifo[channel_num]
+                            .push_back(((word >> 8) & 0b11111111) as i8);
+                        bus.apu.direct_sound_fifo[channel_num]
+                            .push_back(((word >> 16) & 0b11111111) as i8);
+                        bus.apu.direct_sound_fifo[channel_num]
+                            .push_back(((word >> 24) & 0b11111111) as i8);
                     }
                     ChunkSize::Halfword => {
-                        let halfword = bus.read_halfword(self.src_addr);
-                        bus.apu.direct_sound_fifo[channel_num].push_back((halfword & 0b11111111) as i8);
-                        bus.apu.direct_sound_fifo[channel_num].push_back(((halfword >> 8) & 0b11111111) as i8);
+                        let halfword = bus.read_halfword(self.src_addr as usize);
+                        bus.apu.direct_sound_fifo[channel_num]
+                            .push_back((halfword & 0b11111111) as i8);
+                        bus.apu.direct_sound_fifo[channel_num]
+                            .push_back(((halfword >> 8) & 0b11111111) as i8);
                     }
                     _ => {
                         println!("DMA chunk size must be Word or Halfword");
                     }
                 };
-                
-                self.src_addr += self.src_increment * self.chunk_size as usize;
+
+                self.src_addr += self.src_increment * self.chunk_size as u32;
             }
         }
